refactor(frontend): rename Tokens list component
