+++ src/frontend/src/lib/components/tokens/Tokens.svelte
@@ -3,7 +3,7 @@
 	import NetworksSwitcher from '$lib/components/networks/NetworksSwitcher.svelte';
 	import ManageTokensButton from '$lib/components/tokens/ManageTokensButton.svelte';
 	import TokensMenu from '$lib/components/tokens/TokensMenu.svelte';
-	import TokensSignedIn from '$lib/components/tokens/TokensSignedIn.svelte';
+	import TokensList from '$lib/components/tokens/TokensList.svelte';
 	import Header from '$lib/components/ui/Header.svelte';
 	import { testnetsEnabled } from '$lib/derived/settings.derived';
 	import { i18n } from '$lib/stores/i18n.store';
@@ -20,7 +20,7 @@
 		<TokensMenu slot="end" />
 	</Header>
 
-	<TokensSignedIn />
+	<TokensList />
 
 	<div transition:fade class="mb-4 mt-12 flex w-full justify-center sm:w-auto">
 		<ManageTokensButton />
